requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==5.0.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.1.8
ecdsa==0.19.2
exceptiongroup==1.3.1
fastapi==0.128.8
h11==0.16.0
idna==3.11
passlib==1.7.4
pyasn1==0.6.3
pydantic==2.12.5
pydantic_core==2.41.5
python-jose==3.5.0
requests==2.32.5
rsa==4.9.1
six==1.17.0
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.39.0
